Add script which produces topics statistic
